+++ enterprise_data/tests/test_utils.py
@@ -32,7 +32,6 @@ class EnterpriseEnrollmentFactory(factory.django.DjangoModelFactory):
     id = factory.lazy_attribute(lambda x: FAKER.random_int(min=1))
     enterprise_id = factory.lazy_attribute(lambda x: 'ee5e6b3a069a4947bb8dd2dbc323396c')
     lms_user_id = factory.lazy_attribute(lambda x: FAKER.random_int(min=1))
-    enterprise_user_id = factory.lazy_attribute(lambda x: FAKER.random_int(min=1))
     course_id = factory.lazy_attribute(lambda x: FAKER.slug())
     enrollment_created_timestamp = factory.lazy_attribute(lambda x: '2018-01-01')
     user_current_enrollment_mode = factory.lazy_attribute(lambda x: 'verified')